+++ src/com_server/api_builtins.py
@@ -38,7 +38,7 @@ class Builtins:
         - `/send/get_first` (POST): Responds with the first string response from the serial port after sending data, with data and parameters in request; equivalent to `Connection.get_first_response(is_bytes=False, ...)`
         - `/get/wait` (POST): Waits until connection receives string data given in request; different response for success and failure; equivalent to `Connection.wait_for_response(...)`
         - `/send/get` (POST): Continues sending something until connection receives data given in request; different response for success and failure; equivalent to `Connection.send_for_response(...)`
-        - `/connection_state` (GET): Get the properties of the `Connection` object: connected, timeout, send_interval, conn_obj, available, port 
+        - `/connection_state` (GET): Get the properties of the `Connection` object: connected, timeout, send_interval, conn_obj, available, port
         - `/connected` (GET): Indicates if the serial port is currently connected or not
         - `/list_ports` (GET): Lists all available Serial ports
 
@@ -625,12 +625,12 @@ class Builtins:
                 return {"message": "OK"}
 
         return _SendResponse
-    
+
     def _connection_state(_self) -> t.Type[Connection]:
         """
         Responds with an object of the properties of the connection state.
 
-        - `timeout` (float): The timeout of the object, or how much time it will try doing 
+        - `timeout` (float): The timeout of the object, or how much time it will try doing
         something, such as sending data, before breaking out of the program
         - `send_interval` (float): The time the program will wait before allowing something
         to be sent to the serial port again
@@ -642,7 +642,7 @@ class Builtins:
 
         Arguments:
             None
-        
+
         Response:
 
         - `200 OK`:
@@ -657,10 +657,10 @@ class Builtins:
                         "timeout": self.conn.timeout,
                         "send_interval": self.conn.send_interval,
                         "available": self.conn.available,
-                        "port": self.conn.port
-                    }
+                        "port": self.conn.port,
+                    },
                 }
-            
+
         return _ConnectionState
 
     def _connected(_self) -> t.Type[Connection]: